+++ contrib/storage-googlesheets/src/main/java/org/apache/drill/exec/store/googlesheets/GoogleSheetsBatchReader.java
@@ -30,15 +30,15 @@ import org.apache.drill.exec.physical.resultSet.ResultSetLoader;
 import org.apache.drill.exec.physical.resultSet.RowSetLoader;
 import org.apache.drill.exec.record.MaterializedField;
 import org.apache.drill.exec.record.metadata.TupleMetadata;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsBigIntegerColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsBooleanColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsDateColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsFloatColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsIntegerColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsNumericColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsTimeColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsTimestampColumnWriter;
-import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsVarcharColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsBigIntegerColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsBooleanColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsDateColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsFloatColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsIntegerColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsNumericColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsTimeColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsTimestampColumnWriter;
+import org.apache.drill.exec.store.googlesheets.columns.GoogleSheetsColumnWriter.GoogleSheetsVarcharColumnWriter;
 import org.apache.drill.exec.store.googlesheets.utils.GoogleSheetsRangeBuilder;
 import org.apache.drill.exec.store.googlesheets.utils.GoogleSheetsUtils;
 import org.apache.drill.exec.util.Utilities;
