--- contrib/storage-googlesheets/src/main/java/org/apache/drill/exec/store/googlesheets/columns/GoogleSheetsFloatColumnWriter.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.drill.exec.store.googlesheets.columns;
-
-import org.apache.commons.lang3.StringUtils;
-import org.apache.drill.exec.physical.resultSet.RowSetLoader;
-
-public class GoogleSheetsFloatColumnWriter extends GoogleSheetsColumnWriter {
-
-  public GoogleSheetsFloatColumnWriter(RowSetLoader rowWriter, String colName) {
-    super(rowWriter, colName);
-  }
-
-  @Override
-  public void load(Object rawValue) {
-    String stringValue = (String)rawValue;
-    if (StringUtils.isNotEmpty(stringValue)) {
-      float finalValue;
-      try {
-        finalValue = Float.parseFloat(stringValue);
-      } catch (NumberFormatException e) {
-        finalValue = Float.NaN;
-      }
-      columnWriter.setFloat(finalValue);
-    }
-  }
-}
