--- contrib/storage-googlesheets/src/main/java/org/apache/drill/exec/store/googlesheets/columns/GoogleSheetsIntegerColumnWriter.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.drill.exec.store.googlesheets.columns;
-
-import org.apache.commons.lang3.StringUtils;
-import org.apache.drill.exec.physical.resultSet.RowSetLoader;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class GoogleSheetsIntegerColumnWriter extends GoogleSheetsColumnWriter {
-
-  private static final Logger logger = LoggerFactory.getLogger(GoogleSheetsIntegerColumnWriter.class);
-
-  public GoogleSheetsIntegerColumnWriter(RowSetLoader rowWriter, String colName) {
-    super(rowWriter, colName);
-  }
-
-  @Override
-  public void load(Object rawValue) {
-    String stringValue = (String)rawValue;
-    if (StringUtils.isNotEmpty(stringValue)) {
-      int finalValue;
-      try {
-        finalValue = (int) Math.floor(Double.parseDouble(stringValue));
-        columnWriter.setInt(finalValue);
-      } catch (NumberFormatException e) {
-        logger.info("Could not parse {} into integer from Googlesheets.", stringValue);
-      }
-    }
-  }
-}
